feat(freelancer): add schema for freelancer
