+++ data/freelancer.js
@@ -1,4 +1,4 @@
-const freelancer = [ 
+const freelancer = [
     {
         name: 'Souvik Basu',
         skills: ['React', 'Vue', 'Angular'],
@@ -7,6 +7,15 @@ const freelancer = [
         currency: 'INR',
         image: 'https://pbs.twimg.com/profile_images/1439826328493068288/IyDqiBVT_400x400.jpg'
     },
-]
+    {
+        name: 'Souvik Basu',
+        skills: ['React', 'Vue', 'Angular'],
+        experience: 18,
+        rate: 500,
+        currency: 'INR',
+        image: 'https://pbs.twimg.com/profile_images/1439826328493068288/IyDqiBVT_400x400.jpg'
+    }
+] 
+
 
 export default freelancer;
