+++ data/project.js
@@ -1,0 +1,7 @@
+const countries =  {
+    _id: "dhshdshd",
+    title: "React project",
+    description: "adjbsdk jbsds sudb"
+}
+
+export default countries;
