+++ schema/project.js
@@ -1,0 +1,11 @@
+import { gql } from "apollo-server-express";
+
+const typeDefs = gql`
+  type Project {
+    _id: ID!
+    title :String
+    description: String
+  }
+`;
+
+export default typeDefs;
